--- packages/twenty-front/src/modules/activities/calendar/queries/fragments/calendarEventFragment.ts
@@ -1,20 +1,0 @@
-import { gql } from '@apollo/client';
-
-import { attendeeFragment } from '@/activities/calendar/queries/fragments/attendeeFragment';
-
-export const calendarEventFragment = gql`
-  fragment CalendarEventFragment on TimelineCalendarEvent {
-    id
-    title
-    description
-    location
-    startsAt
-    endsAt
-    isFullDay
-    visibility
-    attendees {
-      ...AttendeeFragment
-    }
-  }
-  ${attendeeFragment}
-`;
