+++ packages/twenty-front/src/modules/activities/calendar/queries/fragments/timelineCalendarEventFragment.ts
@@ -1,0 +1,20 @@
+import { gql } from '@apollo/client';
+
+import { timelineCalendarEventAttendeeFragment } from '@/activities/calendar/queries/fragments/timelineCalendarEventAttendeeFragment';
+
+export const timelineCalendarEventFragment = gql`
+  fragment TimelineCalendarEventFragment on TimelineCalendarEvent {
+    id
+    title
+    description
+    location
+    startsAt
+    endsAt
+    isFullDay
+    visibility
+    attendees {
+      ...TimelineCalendarEventAttendeeFragment
+    }
+  }
+  ${timelineCalendarEventAttendeeFragment}
+`;
