+++ packages/twenty-front/src/modules/activities/calendar/queries/fragments/timelineCalendarEventWithTotalFragment.ts
@@ -1,13 +1,13 @@
 import { gql } from '@apollo/client';
 
-import { calendarEventFragment } from '@/activities/calendar/queries/fragments/calendarEventFragment';
+import { timelineCalendarEventFragment } from '@/activities/calendar/queries/fragments/timelineCalendarEventFragment';
 
 export const timelineCalendarEventWithTotalFragment = gql`
   fragment TimelineCalendarEventsWithTotalFragment on TimelineCalendarEventsWithTotal {
     totalNumberOfCalendarEvents
     timelineCalendarEvents {
-      ...CalendarEventFragment
+      ...TimelineCalendarEventFragment
     }
   }
-  ${calendarEventFragment}
+  ${timelineCalendarEventFragment}
 `;
